Game tick now works based on a modulo system

+++ src/tick-manager.h
@@ -1,12 +1,13 @@
 #pragma once
 
+#include <array>
 #include <cstdint>
 
 class TickManager 
 {
   public:
     void incrementTick();
-    uint64_t& getCurrentTick();
+    int& getCurrentTick();
     bool hasHitTick( int tick_to_check );
     void resetTickCounter();
 };
